Tracing log messages with an UUID
This patch adds a logger to the express request object that wraps a winston logger instance. This logger will add a unique id per request to the log message metadata so you can trace all log messages that belongs to the same request. Additionally, the response
will have also a header 'X-Request-Id' so traceability can be made also in the
client.